// backend/company-service/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const CompanyVerificationStatus = {
  PENDING: 'PENDING',
  VERIFIED: 'VERIFIED',
  REJECTED: 'REJECTED'
} as const

export type CompanyVerificationStatus = (typeof CompanyVerificationStatus)[keyof typeof CompanyVerificationStatus]


export const CompanyMemberRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  RECRUITER: 'RECRUITER'
} as const

export type CompanyMemberRole = (typeof CompanyMemberRole)[keyof typeof CompanyMemberRole]


export const CompanyInviteStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  EXPIRED: 'EXPIRED',
  CANCELLED: 'CANCELLED'
} as const

export type CompanyInviteStatus = (typeof CompanyInviteStatus)[keyof typeof CompanyInviteStatus]